Feature/sprint4/takeover

+++ micro-api/controller/datatransfer/importexport/import_export_api.go
@@ -78,7 +78,7 @@ func ExportData(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, controller.Fail(http.StatusInternalServerError, err.Error()))
 	} else {
 		status = respDTO.GetRespStatus()
-		if common.TIEM_SUCCESS == status.GetCode() {
+		if int32(common.TIEM_SUCCESS) == status.GetCode() {
 			result := controller.BuildCommonResult(int(status.Code), status.Message, DataExportResp{
 				RecordId: respDTO.GetRecordId(),
 			})
@@ -133,7 +133,7 @@ func ImportData(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, controller.Fail(http.StatusInternalServerError, err.Error()))
 	} else {
 		status = respDTO.GetRespStatus()
-		if common.TIEM_SUCCESS == status.GetCode() {
+		if int32(common.TIEM_SUCCESS) == status.GetCode() {
 			result := controller.BuildCommonResult(int(status.Code), status.Message, DataImportResp{
 				RecordId: respDTO.GetRecordId(),
 			})
@@ -183,7 +183,7 @@ func DescribeDataTransport(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, controller.Fail(http.StatusInternalServerError, err.Error()))
 	} else {
 		status = respDTO.GetRespStatus()
-		if common.TIEM_SUCCESS == status.GetCode() {
+		if int32(common.TIEM_SUCCESS) == status.GetCode() {
 			data := &DataTransportRecordQueryResp{
 				TransportRecords: make([]*DataTransportInfo, len(respDTO.GetTransportInfos())),
 			}
